[CODE TEST TWO] Find Bug and Code Review
There is a bug in the program, at least 1 that we know of.  Let us know
when / if you find it and how you approached debugging and the steps
taken.

Also please include code review comments regarding style of code,
data structures / algorithms used, general problem solving method or any
other remarks in your submission.

To submit, push your branch (not named master) to the repo and do a pull
request.